+++ src/test/java/org/torproject/onionoo/server/ResourceServletTest.java
@@ -153,7 +153,7 @@ public class ResourceServletTest {
             "0025C136C1F3A9EEFE2AE3F918F03BFA21B5070B" })),
         new TreeSet<>(Arrays.asList(
         new String[] { "001C13B3A55A71B977CA65EC85539D79C653A3FC" })),
-        "0.2.3.25", "ppp-62-216-201-221.dynamic.mnet-online.de",
+        "0.2.3.25", "linux", "ppp-62-216-201-221.dynamic.mnet-online.de",
         Arrays.asList(
             new String[] { "ppp-62-216-201-221.dynamic.mnet-online.de" }),
         null, true);
@@ -170,7 +170,7 @@ public class ResourceServletTest {
         new TreeSet<String>(Arrays.asList(new String[] {
             "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })),
         new TreeSet<>(Arrays.asList(new String[] {
-            "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })), null,
+            "000C5F55BD4814B917CC474BD537F1A3B33CCE2A" })), null, null,
             "c-68-38-171-200.hsd1.in.comcast.net",
         Arrays.asList(
             new String[] {"c-68-38-171-200.hsd1.in.comcast.net"}),
@@ -187,8 +187,8 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-16 18:00:00"), "AS6830",
         "1024d/51e2a1c7 \"steven j. murdoch\" "
         + "<tor+steven.murdoch@cl.cam.ac.uk> <fb-token:5sr_k_zs2wm=>",
-        new TreeSet<String>(), new TreeSet<String>(), "0.2.3.24-rc-dev", null,
-        null, null, false);
+        new TreeSet<String>(), new TreeSet<String>(), "0.2.3.24-rc-dev",
+        "windows xp", null, null, null, false);
     this.relays.put("0025C136C1F3A9EEFE2AE3F918F03BFA21B5070B",
         relayTimMayTribute);
     this.bridges = new TreeMap<>();
@@ -199,7 +199,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-21 18:07:03"), false,
         new TreeSet<>(Arrays.asList(new String[] { "Valid" })), -1L,
         null, DateTimeHelper.parse("2013-04-20 15:37:04"), null, null,
-        null, null, "0.2.2.39", null, null, null, true);
+        null, null, "0.2.2.39", null, null, null, null, true);
     this.bridges.put("0000831B236DFF73D409AD17B40E2A728A53994F",
         bridgeec2bridgercc7f31fe);
     org.torproject.onionoo.docs.SummaryDocument bridgeUnnamed =
@@ -209,7 +209,7 @@ public class ResourceServletTest {
         DateTimeHelper.parse("2013-04-20 17:37:04"), false,
         new TreeSet<>(Arrays.asList(new String[] { "Valid" })), -1L,
         null, DateTimeHelper.parse("2013-04-14 07:07:05"), null, null,
-        null, null, null, null, null, null, null);
+        null, null, null, null, null, null, null, null);
     this.bridges.put("0002D9BDBBC230BD9C78FF502A16E0033EF87E0C",
         bridgeUnnamed);
     org.torproject.onionoo.docs.SummaryDocument bridgegummy =
@@ -220,7 +220,7 @@ public class ResourceServletTest {
         new TreeSet<>(Arrays.asList(new String[] { "Running",
             "Valid" })), -1L, null,
         DateTimeHelper.parse("2013-01-16 21:07:04"), null, null, null,
-        null, "0.2.4.4-alpha-dev", null, null, null, false);
+        null, "0.2.4.4-alpha-dev", "windows 7", null, null, null, false);
     this.bridges.put("1FEDE50ED8DBA1DD9F9165F78C8131E4A44AB756",
         bridgegummy);
   }
@@ -1646,6 +1646,44 @@ public class ResourceServletTest {
   public void testVersionStart() {
     /* This is also correct when comparing strings. */
     this.assertErrorStatusCode("/summary?version=*", 400);
+  }
+
+  @Test(timeout = 100)
+  public void testOperatingSystemLinux() {
+    this.assertSummaryDocument(
+        "/summary?os=linux", 1, new String[] {"TorkaZ"}, 0, null);
+  }
+
+  @Test(timeout = 100)
+  public void testOperatingSystemLinuxMixedCaps() {
+    this.assertSummaryDocument(
+        "/summary?os=LiNUx", 1, new String[] {"TorkaZ"}, 0, null);
+  }
+
+  @Test(timeout = 100)
+  public void testOperatingSystemLin() {
+    this.assertSummaryDocument(
+        "/summary?os=lin", 1, new String[] {"TorkaZ"}, 0, null);
+  }
+
+  @Test(timeout = 100)
+  public void testOperatingSystemWindows() {
+    this.assertSummaryDocument(
+        "/summary?os=windows", 1, new String[] {"TimMayTribute"},
+        1, new String[] {"gummy"});
+  }
+
+  @Test(timeout = 100)
+  public void testOperatingSystemWindowsExperience() {
+    this.assertSummaryDocument(
+        "/summary?os=windows xp", 1, new String[] {"TimMayTribute"},
+        0, null);
+  }
+
+  @Test(timeout = 100)
+  public void testOperatingSystemWindows7() {
+    this.assertSummaryDocument(
+        "/summary?os=windows 7", 0, null, 1, new String[] {"gummy"});
   }
 
   @Test